refactor(training): extract MLX model to model.py (behavior-preserving)
train.py had accumulated the MLX model architecture, the training loop, the
3-backend dispatch, and the CLI/subprocess entry, repeatedly pushing it against
the 800-line module-size limit. Extract the stable, serving-shared model layer
into a new training/autoresearch/model.py:

- ModelConfig, the MLX model components (RMSNorm/RoPE/Attention/FeedForward/
  TransformerBlock/GPTModel), compute_loss, the HAS_MLX-else stubs, and checkpoint
  I/O (save_checkpoint / load_checkpoint / save_inference_bundle) move verbatim.
- train.py re-exports every name (`from ...model import ...`), so existing
  `from ...train import GPTModel/ModelConfig/compute_loss/...` paths in cuda.py,
  tests, and elsewhere keep working unchanged.
- MLXProvider now imports the model directly from model.py: the serving path no
  longer reaches into the training-loop module (fixes a backwards dependency).

No behavior change. train.py drops from 799 to 487 lines, giving durable runway
for the remaining pipeline work. The one test that patched
`train.save_checkpoint` (exercised via save_inference_bundle, now in model.py) is
repointed to `model.save_checkpoint`.
