+++ CHANGELOG.md
@@ -17,6 +17,10 @@ All notable changes to this project will be documented in this file.
   - Reward-weighted loss (reward-weighted regression): `--loss-weight-by-score` (`uniform` | `linear` | `softmax`) scales each training example's loss by a weight derived from its score, the soft counterpart to elite filtering (keep every record, but lean the gradient toward high-reward constructions). The weight is applied **per example** (each example's mean completion loss is weighted, then averaged), returned as a separate per-example weight vector rather than folded into the per-token mask, so a long completion cannot drown out a short high-reward one. Non-uniform modes are mean-normalized to 1.0 (they change relative emphasis, not the overall step size); `--loss-weight-temperature` (must be `> 0` for `softmax`) is a continuous knob between uniform and soft-elite; no score spread falls back to uniform. `uniform` (default) is byte-identical to prior training. Applies to the `mlx` and `cuda` backends (the `mlxlm` backend rejects non-uniform modes); the setting is recorded in the registry `data_stats`. Pure helper `score_loss_weights` lives in `sequence_format.py`. See `docs/mlx-training.md`.
   - Self-improving loop (ReST-EM / expert iteration): new `autoctx self-improve` command runs an outer training loop. Each round trains on the current dataset, samples constructions from the trained model and scores them in-scenario, keeps the top `--elite-fraction` by score, and appends them as new training records, so the model generates its own training data biased toward the best of what it can produce (the iterated form of rejection-sampling fine-tuning). Because a round trains before appending its own elite, the loop runs one final training pass over the full accumulated dataset (into `<output-dir>/final`, reported as `final_avg_score`) so the shipped model reflects every collected sample. Generated elite records inherit the seed dataset's representative `context` so every training example shares the same context prefix instead of mixing rich seed context with empty prefixes. Sampling uses a positive temperature so collected samples are diverse. Carries `--score-conditioned` through every round and prints a per-round table (avg_score, samples, elite, dataset size); `--json` returns the structured `history` plus `final_model_dir` / `final_avg_score`. Loop counts (`--rounds` / `--samples-per-round` / `--train-steps`) are validated as positive integers. Sample collection is MLX-only (`run_training(collect_samples_path=...)`); the cuda/mlxlm backends reject it with a clear `NotImplementedError`. Pure helpers (`select_elite_samples`, `samples_to_records`, `representative_context`) live in `training/autoresearch/self_improve.py`. See `docs/mlx-training.md`.
 
+### Changed
+
+- The baseline MLX GPT model (`ModelConfig`, `GPTModel` + components, `compute_loss`, and checkpoint I/O `save_checkpoint` / `load_checkpoint` / `save_inference_bundle`) was extracted from `training/autoresearch/train.py` into a new `training/autoresearch/model.py`, separating the (serving-shared, stable) model architecture from the training-loop orchestration. Behavior is unchanged: every name is re-exported from `train.py` so existing `from ...train import` paths keep working, and `MLXProvider` now imports the model directly from `model.py` (the serving path no longer reaches into the training-loop module). This keeps `train.py` within the 800-line module-size budget.
+
 ### Fixed
 
 - autoresearch assessment generation collapsed on structured-output scenarios. Greedy argmax with no per-sample seed produced N identical completions, and for JSON-style strategies it decoded into the most frequent special token, yielding `valid_rate=0`. A shared, pure `generation_logit_mask_values` now blocks both the undecodable phantom-id gap (when the BPE learns fewer merges than `base_vocab_size`, sampling those ids previously raised `KeyError: Invalid token for decoding`) and the structural `<|scenario|>` / `<|context|>` / `<|strategy|>` tokens. The mask is applied in both the greedy and sampling paths and is reused by `MLXProvider` inference so saved bundles cannot sample an undecodable id at serving time.
